Cleaned up debug code; Moved early exit from collision detection code into helpers file, and renamed the function.
