+++ src/app.js
@@ -4,6 +4,7 @@ const config = {
   cellSize: 1,
   renderTick: 10,
   playTick: 10,
+  runtimeDebug: false, // This should be false in every commit, only change during dev.
 }
 
 const elements = {
@@ -13,4 +14,6 @@ const elements = {
 const game = new Game(config, elements);
 game.init();
 
-window.game = game; // runtime hackery
+if (config.runtimeDebug) {
+  window.game = game; // Allows developer to examine the game object at runtime.
+}
